+++ src/debugger/evaluation/primitivetypes/types.h
@@ -24,7 +24,7 @@ using PrimitiveValue = std::variant<std::monostate, bool, uint8_t, int8_t, uint1
                                     uint32_t, int32_t, uint64_t, int64_t, float, double, WCHAR>;
 
 bool IsPrimitiveType(CorElementType elemType);
-HRESULT GetOperandData(ICorDebugValue *pValue, CorElementType elemType, PrimitiveValue &primValue);
+HRESULT GetPrimitiveData(ICorDebugValue *pValue, PrimitiveValue &primValue);
 CorElementType GetCorElementType(const PrimitiveValue &primValue);
 std::string_view GetManagedTypeName(const PrimitiveValue &primValue);
 std::string ToString(const PrimitiveValue &primValue);
@@ -36,6 +36,8 @@ HRESULT CalculateBinary(Parser::SyntaxKind kind, const PrimitiveValue &leftValue
                         PrimitiveValue &outputValue, std::string &output);
 
 HRESULT ForceCastToUint(ICorDebugValue *pInputValue, uint32_t &number);
+HRESULT ImplicitCastIntLiteral(ICorDebugValue *pSrcValue, ICorDebugValue *pDstValue);
+HRESULT ImplicitCast(ICorDebugValue *pSrcValue, ICorDebugValue *pDstValue);
 
 } // namespace dncdbg::PrimitiveTypes